fix test localization
