+++ napari/utils/_magicgui.py
@@ -8,6 +8,8 @@ their custom classes or types should be converted into GUIs.  Then, when the
 end-user annotates one of their function arguments with a type hint using one
 of those custom classes, magicgui will know what to do with it.
 
+Because of headless tests the tests for this module are
+in napari/_tests/test_magicgui.py
 """
 from __future__ import annotations
